+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 All Notable changes to **Digipolis DG Cookiebot** module.
 
+## [Unreleased]
+
+### Added
+
+* CCM-61: Add interface to set the cookie category manually for Drupal
+  javascript libraries.
+
+### Changed
+
+* CCM-61: Change minimal Drupal version to 10.1.
+
 ## [1.3.0]
 
 ### Added